Refactor configuration files
