+++ errai-config/src/main/java/org/jboss/errai/config/ErraiAppConfiguration.java
@@ -16,11 +16,19 @@
 
 package org.jboss.errai.config;
 
+import org.jboss.errai.config.ErraiSeparateAppConfiguration.Ioc;
+import org.jboss.errai.config.ErraiSeparateAppConfiguration.Marshalling;
+
+import java.util.Optional;
+
+import static org.jboss.errai.config.ErraiSeparateAppConfiguration.Others;
+
 /**
  * @author Tiago Bento <tfernand@redhat.com>
  */
-public interface ErraiAppConfiguration extends ErraiSeparateAppConfiguration.Ioc, ErraiSeparateAppConfiguration.Others {
+public interface ErraiAppConfiguration extends Ioc, Marshalling, Others {
 
   boolean isAptEnvironment();
 
+  Optional<String> custom(final String propertyName);
 }
