+++ errai-config/src/main/java/org/jboss/errai/config/ErraiModulesConfiguration.java
@@ -16,13 +16,12 @@
 
 package org.jboss.errai.config;
 
-import org.jboss.errai.codegen.meta.MetaClass;
-
-import java.util.Set;
+import static org.jboss.errai.config.ErraiSeparateModuleConfiguration.DataBinding;
+import static org.jboss.errai.config.ErraiSeparateModuleConfiguration.Ioc;
+import static org.jboss.errai.config.ErraiSeparateModuleConfiguration.Marshalling;
 
 /**
  * @author Tiago Bento <tfernand@redhat.com>
  */
-public interface ErraiModulesConfiguration extends ErraiSeparateModuleConfiguration.Ioc,
-        ErraiSeparateModuleConfiguration.DataBinding, ErraiSeparateModuleConfiguration.Marshalling {
+public interface ErraiModulesConfiguration extends Ioc, DataBinding, Marshalling {
 }
