+++ errai-config/src/main/java/org/jboss/errai/config/propertiesfile/PropertiesUtil.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.jboss.errai.config;
+package org.jboss.errai.config.propertiesfile;
 
 import org.jboss.errai.common.metadata.ScannerSingleton;
