+++ errai-config/src/main/java/org/jboss/errai/config/rebind/EnvUtil.java
@@ -16,51 +16,10 @@
 
 package org.jboss.errai.config.rebind;
 
-import org.jboss.errai.common.metadata.ErraiAppPropertiesFiles;
-import org.jboss.errai.common.rebind.CacheStore;
-import org.jboss.errai.common.rebind.CacheUtil;
-import org.jboss.errai.config.ErraiAppPropertiesErraiModulesConfiguration;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.URL;
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.PropertyResourceBundle;
-import java.util.ResourceBundle;
-import java.util.concurrent.ConcurrentHashMap;
-
 /**
  * @author Mike Brock
  */
 public abstract class EnvUtil {
-
-  public static class EnvironmentConfigCache implements CacheStore {
-    private volatile EnvironmentConfig environmentConfig;
-    private final Map<String, String> permanentProperties = new ConcurrentHashMap<>();
-
-    public EnvironmentConfigCache() {
-      clear();
-    }
-
-    @Override
-    public synchronized void clear() {
-      environmentConfig = newEnvironmentConfig();
-      environmentConfig.getFrameworkProperties().putAll(permanentProperties);
-    }
-
-    public synchronized EnvironmentConfig get() {
-      return environmentConfig;
-    }
-
-    public void addPermanentFrameworkProperty(final String name, final String value) {
-      permanentProperties.put(name, value);
-      environmentConfig.getFrameworkProperties().put(name, value);
-    }
-  }
 
   private static volatile Boolean _isJUnitTest;
 
@@ -109,74 +68,4 @@ public abstract class EnvUtil {
     isProdMode();
   }
 
-  private static Logger log = LoggerFactory.getLogger(EnvUtil.class);
-
-  public static Collection<URL> getErraiAppPropertiesFilesUrls() {
-    final ClassLoader contextClassLoader = Thread.currentThread().getContextClassLoader();
-    final ClassLoader envUtilClassLoader = EnvUtil.class.getClassLoader();
-    return ErraiAppPropertiesFiles.getUrls(contextClassLoader, envUtilClassLoader);
-  }
-
-  public static void clearCache() {
-    CacheUtil.getCache(EnvironmentConfigCache.class).clear();
-  }
-
-  /**
-   * @return an instance of {@link EnvironmentConfig}. Do NOT retain a reference to this value. Call every time
-   * you need additional configuration information.
-   */
-  public static EnvironmentConfig getEnvironmentConfig() {
-    return CacheUtil.getCache(EnvironmentConfigCache.class).get();
-  }
-
-  private static EnvironmentConfig newEnvironmentConfig() {
-    final Map<String, String> frameworkProps = new HashMap<>();
-
-    for (final URL url : EnvUtil.getErraiAppPropertiesFilesUrls()) {
-      InputStream inputStream = null;
-      try {
-        log.debug("checking " + url.getFile() + " for configured types ...");
-        inputStream = url.openStream();
-
-        final ResourceBundle props = new PropertyResourceBundle(inputStream);
-        for (final String key : props.keySet()) {
-          final String value = props.getString(key);
-          if (frameworkProps.containsKey(key)) {
-            if (isListValuedProperty(key)) {
-              // TODO should validate that different values don't conflict
-              final String oldValue = frameworkProps.get(key);
-              final String newValue = oldValue + " " + value;
-              log.debug("Merging property {} = {}", key, newValue);
-              frameworkProps.put(key, newValue);
-            } else {
-              log.warn("The property {} has been set multiple times.", key);
-              frameworkProps.put(key, value);
-            }
-          } else {
-            frameworkProps.put(key, value);
-          }
-        }
-      } catch (final IOException e) {
-        throw new RuntimeException("error reading ErraiApp.properties", e);
-      } finally {
-        if (inputStream != null) {
-          try {
-            inputStream.close();
-          } catch (final IOException e) {
-            //
-          }
-        }
-      }
-    }
-
-    return new EnvironmentConfig(frameworkProps);
-  }
-
-  private static boolean isListValuedProperty(final String key) {
-    return key.equals(ErraiAppPropertiesErraiModulesConfiguration.IOC_ENABLED_ALTERNATIVES) || key.equals(
-            ErraiAppPropertiesErraiModulesConfiguration.BINDABLE_TYPES) || key.equals(
-            ErraiAppPropertiesErraiModulesConfiguration.SERIALIZABLE_TYPES) || key.equals(
-            ErraiAppPropertiesErraiModulesConfiguration.NON_SERIALIZABLE_TYPES) || key.equals(
-            ErraiAppPropertiesErraiModulesConfiguration.MAPPING_ALIASES);
-  }
 }
